requirements.txt
# Google API dependencies
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6

# Image processing
Pillow>=9.0.0
requests>=2.26.0

# AI/ML dependencies
google-generativeai>=0.1.0
openai>=1.0.0

# Utility packages
python-dotenv>=0.19.0
tqdm>=4.65.0
loguru>=0.7.0

# Web related
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.0